Use a less garbage regex
